backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pydantic-core==2.16.2
pydantic==2.6.1
email-validator==2.1.0.post1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
python-multipart==0.0.6
langchain-core==0.1.18
openai==1.8.0
numpy==2.2.0
websockets==12.0
aiosqlite==0.19.0
alembic==1.13.1
bcrypt==4.1.2
httpx==0.26.0 